The basic logic is implemented

+++ authenticationServer/src/main/java/pl/university/authenticationserver/user/service/UserService.java
@@ -2,14 +2,19 @@ package pl.university.authenticationserver.user.service;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import pl.university.authenticationserver.user.document.User;
+import pl.university.authenticationserver.user.dto.CreateUserDTO;
 import pl.university.authenticationserver.user.dto.GetUserDTO;
 import pl.university.authenticationserver.user.dto.UpdateUserDTO;
 import pl.university.authenticationserver.user.exceptions.ApiRequestException;
 import pl.university.authenticationserver.user.repository.UserRepository;
+import pl.university.authenticationserver.user.utils.PasswordEncoder;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -18,6 +23,24 @@ import java.util.stream.Collectors;
 public class UserService {
 
     private final UserRepository userRepository;
+
+
+    public void createUser(CreateUserDTO dto) {
+
+        String login = dto.getLogin();
+        if (this.getByLogin(login).isPresent()) {
+           throw new ApiRequestException("User already exists");
+        }
+
+        User newUser = new User(
+                login,
+                PasswordEncoder.encode(dto.getPassword()),
+                dto.getFirstName(),
+                dto.getLastName()
+        );
+
+        userRepository.insert(newUser);
+    }
 
 
     public List<GetUserDTO> getUsers() {
@@ -35,34 +58,15 @@ public class UserService {
     }
 
 
-
-
-
-
-
-
-
-
-    // todo сделать валидацию и поменить на user
-    public Optional<User> getByLogin(@NonNull String login) {
-        return userRepository.findByLogin(login).stream().filter(user -> login.equals(user.getLogin())).findFirst();
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
     public void updateUser(UpdateUserDTO dto, String id) {
+        //-----------------------------
+        // user can only update himself.
+        //-----------------------------
         userRepository.findById(id)
                 .map(user -> {
+                    if (!Objects.equals(user.getLogin(), this.getAuthUserLogin())) {
+                        throw new ApiRequestException("User is not authorized to update this profile");
+                    }
                     user.setFirstName(dto.getFirstName());
                     user.setLastName(dto.getLastName());
                     return userRepository.save(user);
@@ -72,8 +76,26 @@ public class UserService {
 
 
     public void deleteUser(String id) {
+        //-----------------------------
+        // user can only delete himself.
+        //-----------------------------
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new ApiRequestException("User not found for id: " + id));
-        userRepository.delete(user);
+
+       if (!Objects.equals(user.getLogin(), this.getAuthUserLogin()))
+           throw new ApiRequestException("User is not authorized to delete this profile");
+
+       userRepository.delete(user);
+    }
+
+
+    public Optional<User> getByLogin(@NonNull String login) {
+        return userRepository.findByLogin(login).stream().filter(user -> login.equals(user.getLogin())).findFirst();
+    }
+
+
+    public String getAuthUserLogin() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return (String) authentication.getPrincipal();
     }
 }
